Enhance Post modal components for auth handling
- Improve PostViewModal component with user authentication status passed as props to sub components for better user auth handling and readability.
- Enhance auth handling and rendering logic for better performance and code consistency in PostExpandedView component.
